+++ src/presentational-components/shared/filter-select.js
@@ -1,7 +1,8 @@
-import React from 'react';
+import React, { useState, memo } from 'react';
 import PropTypes from 'prop-types';
 import { InternalSelect } from '@data-driven-forms/pf4-component-mapper/dist/cjs/select';
 import { FilterIcon } from '@patternfly/react-icons';
+import isEqual from 'lodash/isEqual';
 
 const ValueContainer = ({ children }) => (
   <div className="filter-value-container" style={{ position: 'absolute' }}>
@@ -17,30 +18,19 @@ ValueContainer.propTypes = {
   ]).isRequired
 };
 
-const FilterSelect = (props) => {
-  if (!props.options || props.options.length === 0) {
-    return (
-      <div
-        key="filter-select-placeholder"
-        id="filter-select-placeholder"
-        className="filter-select"
-      >
-        <InternalSelect
-          components={{ ValueContainer }}
-          simpleValue={false}
-          options={[]}
-          isDisabled
-          onChange={() => {}}
-        />
-      </div>
-    );
-  }
-
+const FilterSelect = ({ onChange, ...props }) => {
+  const [stateValue, setValue] = useState(undefined);
   return (
     <div key="filter-select" id="filter-select" className="filter-select">
       <InternalSelect
-        components={{ ValueContainer }}
+        isDisabled={!props.options || props.options.length === 0}
+        name="filter-select"
         simpleValue={false}
+        onChange={(value) => {
+          onChange(value || stateValue);
+          setValue(value || stateValue);
+        }}
+        value={stateValue}
         {...props}
       />
     </div>
@@ -48,7 +38,10 @@ const FilterSelect = (props) => {
 };
 
 FilterSelect.propTypes = {
-  options: PropTypes.array
+  options: PropTypes.array,
+  onChange: PropTypes.func.isRequired
 };
 
-export default FilterSelect;
+export default memo(FilterSelect, (prevProps, nextProps) =>
+  isEqual(prevProps.options, nextProps.options)
+);
